Feedback collectors return values changed to int.

+++ nsgp/sampling/FeedbackCollector.py
@@ -7,5 +7,5 @@ from genepro.node import Node
 class FeedbackCollector(ABC):
 
     @abstractmethod
-    def collect_feedback(self, pairs: List[Tuple[Node, Node]]) -> List[float]:
+    def collect_feedback(self, pairs: List[Tuple[Node, Node]]) -> List[int]:
         pass
